logout toast update
