+++ frontend/src/pages/Admin/Admin.jsx
@@ -13,21 +13,12 @@ import { Manage } from "./Components/Setting/Manage";
 import { AddBroadcasts } from "./Components/Broadcast/AddBroadcasts";
 import { AddFaq } from "./Components/Faq/AddFaq";
 import { logout } from "../../store/actions/auth";
-import { SimpleToast } from "../../components/util/Toast/index.js";
 
 import { useDispatch } from "react-redux";
 
 export const Admin = () => {
   const [tab, setTab] = useState(1);
   const dispatch = useDispatch();
-  const [open, setOpenToast] = React.useState(false);
-
-  const handleCloseToast = (event, reason) => {
-    if (reason === "clickaway") {
-      return;
-    }
-    setOpenToast(false);
-  };
   return (
     <div style={{ minHeight: "100vh" }}>
       <div className={style["dashing"]}>
@@ -99,12 +90,7 @@ export const Admin = () => {
           </div>
           <div
             className={style["features-icons1"]}
-            onClick={() => {
-              setOpenToast(true);
-              setTimeout(() => {
-                logout(dispatch);
-              }, 3000);
-            }}
+            onClick={() => logout(dispatch)}
           >
             <i
               className="fas fa-sign-out-alt fa-fw fa-lg"
@@ -139,11 +125,6 @@ export const Admin = () => {
           ) : null}
         </div>
       </div>
-      <SimpleToast
-        open={open}
-        message="Successfully Logged Out!"
-        handleCloseToast={handleCloseToast}
-      />
     </div>
   );
 };
