+++ frontend/src/store/actions/auth.js
@@ -1,5 +1,6 @@
 export const logout = (dispatch) => {
   localStorage.removeItem("token");
   localStorage.removeItem("isSuperAdmin");
+  localStorage.setItem("log", true);
   window.location = "/admin";
 };
