+++ frontend/src/store/reducers/reducer.js
@@ -4,6 +4,7 @@ let initialState = {
   //auth
   token: localStorage.getItem("token") || null,
   isSuperAdmin: localStorage.getItem("isSuperAdmin") || false,
+  logout: localStorage.getItem("log") || false,
 };
 
 const reducer = (state = initialState, action) => {
